+++ controller/user.py
@@ -1,0 +1,9 @@
+from flask import Blueprint
+from repository.UserRepository import UserRepository
+
+userRepository =  UserRepository()
+user_bp = Blueprint("user", __name__)
+
+@user_bp.route("/api/students/exist/<studentNo>")
+def checkStudentNo(studentNo: str):
+    return {"exists" : userRepository.existByStudentNo(studentNo=studentNo)}
