+++ repository/UserRepository.py
@@ -1,6 +1,5 @@
 from pymongo import MongoClient
 
-
 class UserRepository:
     def __init__(self):
         self.__db = MongoClient("127.0.0.1").conbrief.user
@@ -16,7 +15,6 @@ class UserRepository:
     def existByStudentNo(self, studentNo: str) -> bool:
         return True if self.__db.find_one({"studentNo": studentNo}) else False
     
-
     def findByStudentNo(self, studentNo: str) -> dict:
         student = self.__db.find_one({"studentNo": studentNo})
         return student
